moving pagination into card list component
