+++ web/src/components/studies/IGOpsStudies.module.css
@@ -11,17 +11,6 @@
   margin-bottom: 2rem;
 }
 
-.pagination-container {
-  margin-top: 2rem;
-
-  small {
-    display: block;
-    margin-top: 0.5rem;
-    color: #6c757d;
-    text-align: center;
-  }
-}
-
 @media (max-width: 768px) {
   .no-studies-message {
     padding: 2rem 1rem;
